+++ CHANGES.md
@@ -6,6 +6,8 @@ Changes 0.8.0-prerelease:
 
 * enhancement: add Dilithium ARMv7 support
 
+* enhancement: add Kyber ARMv7 support
+
 Changes 0.7.0:
 
 * enhancement: add XDRBG256 - the SHAKE256-based DRNG discussed for SP800-90A